# encoding/xml: make use of reflect.TypeAssert